crop-assist/requirements.txt
flask==2.3.3
joblib==1.3.2
groq==0.4.2
python-dotenv==1.0.0
requests==2.31.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.3.2
transformers==4.41.2
torch==2.2.0
